+++ src/fights/competitor.clj
@@ -17,28 +17,24 @@
   [id]
   (get-ijf "competitor.info" id))
 
-(comment
-  (def hiraoka (info 30))
-  hiraoka)
-
 (defn wrl-current
   [id]
   (get-ijf "competitor.wrl_current" id))
 
-(comment
-  (wrl-current 30))
-
 (defn fights-statistics
   [id]
   (get-ijf "competitor.fights_statistics" id))
-
-(comment
-  (fights-statistics 30))
 
 (defn contests
   [id]
   (-> (get-ijf "competitor.contests" id)
       :contests))
 
-(comment
-  (count (contests 30)))
+(defn id-fight
+  [id]
+  (->> (contests id)
+       (map :id_fight)))
+
+(defn contests-statistics
+ [id]
+ (get-ijf "competitor.contests_statistics" id))