+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.2.0-SNAPSHOT
+- src/fights/db.clj create tables. `persons`, `contests` and `competitions`.
+- test/fights/db_test.clj
+- src/fights/competitor.clj
+- src/fights/core.clj
+
 
 ## 0.1.0 - 2022-08-12
 - project (re)started.